windows/requirements.txt
# OBS Tutorial Recorder - Windows
# Python 3.11+ required

# Core
websockets>=11.0

# System Tray
pystray>=0.19.4
Pillow>=9.0.0

# UI
customtkinter>=5.2.0

# Device Enumeration
pywin32>=306

# Process Management
psutil>=5.9.0

# Notifications
plyer>=2.1.0

# Transcription (optional)
# faster-whisper>=0.9.0

# Packaging (dev only)
# pyinstaller>=6.0.0
